+++ packages/reactabular-table/src/body.js
@@ -27,28 +27,16 @@ class Body extends React.Component {
     return React.createElement(
       components.body.wrapper,
       props,
-      rows.map((r, i) => {
-        if (process.env.NODE_ENV !== 'production') {
-          // Arrays cannot have rowKeys by definition so we have to go by index there.
-          if (!Array.isArray(r) && !{}.hasOwnProperty.call(r, rowKey)) {
-            console.warn( // eslint-disable-line no-console
-              'Table.Body - Missing valid rowKey!',
-              r,
-              rowKey
-            );
-          }
-        }
-
-        return React.createElement(BodyRow, {
-          key: `${r[rowKey] || i}-row`,
+      rows.map((rowData, rowIndex) => (
+        React.createElement(BodyRow, {
+          key: resolveRowKey({ rowData, rowIndex, rowKey }),
           components: components.body,
           onRow,
-          row: r,
-          rowIndex: i,
-          rowData: rows[i],
+          rowIndex,
+          rowData,
           columns: bodyColumns
-        });
-      })
+        })
+      ))
     );
   }
   getRef() {
@@ -60,6 +48,23 @@ Body.defaultProps = {
   onRow: () => {}
 };
 Body.contextTypes = tableBodyContextTypes;
+
+function resolveRowKey({ rowData, rowIndex, rowKey }) {
+  if (typeof rowKey === 'function') {
+    return `${rowKey({ rowData, rowIndex })}-row`;
+  } else if (process.env.NODE_ENV !== 'production') {
+    // Arrays cannot have rowKeys by definition so we have to go by index there.
+    if (!Array.isArray(rowData) && !{}.hasOwnProperty.call(rowData, rowKey)) {
+      console.warn( // eslint-disable-line no-console
+        'Table.Body - Missing valid rowKey!',
+        rowData,
+        rowKey
+      );
+    }
+  }
+
+  return `${rowData[rowKey] || rowIndex}-row`;
+}
 
 function omitOnRow(props) {
   const { onRow, ...ret } = props; // eslint-disable-line no-unused-vars